refactor: `<UserVideo>`를 `<UserVideoGrid>` 내부로 이동
- `<UserVideo>`가 하위 컴포넌트이기 때문
